# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API Docs written by json. This is Docs for Assigment 6",
        "version": "1.0.0",
        "title": "Travellog API",
        "contact": {
            "email": "17520279@gm.uit.edu.vn",
            "name": "Duong Le Thanh Binh"
        },
        "servers": [
            "http://localhost:4000"
        ]
    },
    "paths": {
...
        "/blogs": {
            "get": {
                "tags": [
                    "/blogs"
                ],
                "summary": "Get response data from blogs page",
                "description": "Get response data from blogs page. Include: <br/><ul><li>_id</li><li>name</li><li>title</li><li>content</li><li>__v</li></ul>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get all json data"
                    }
                }
            },
            "post": {
                "tags": [
                    "/blogs"
                ],
                "parameters": [
                    {
                        "name": "blogs",
                        "in": "body",
                        "description": "blogs that we want to create",
                        "schema": {
                            "$ref": "#/definitions/blogs"
                        }
                    }
                ],
                "summary": "Post to mongodb atlas, the get response.",
                "description": "Post to mongodb atlas, the get response.<br/>In other word, add new blog<br/> Example for 'name' must be: <br/><ul><li>Ha Noi</li><li>Ho Chi Minh</li><li>Hue</li><li>Da Nang</li><li>Phong Nha</li><li>Sapa</li><li>Ben Tre</li></ul>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get back json data"
                    }
                }
            }
        },
        "/blogs/{blogsID}": {
            "get": {
                "tags": [
                    "/blogs"
                ],
                "parameters": [
                    {
                        "name": "blogsID",
                        "in": "path",
                        "description": "ID of blogs that needs to be search. Example: <ul><li>Ha Noi</li><li>Ho Chi Minh</li><li>Hue</li><li>Da Nang</li><li>Phong Nha</li><li>Sapa</li><li>Ben Tre</li></ul>",
                        "required": true,
                        "type": "string"
                    }
                ],
                "summary": "Search by ID from blogs page",
                "description": "Get response from search ID blogs page<br/> Include: <br/><ul><li>_id</li><li>name</li><li>title</li><li>content</li><li>__v</li></ul>",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Get all search json data"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "blogs": {
            "required": [
                "name",
                "title",
                "content"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                }
            }
        },
...
    }
}